move stuff outside of utils.__init__
