+++ sklearn/datasets/_arff_parser.py
@@ -10,11 +10,8 @@ import scipy as sp
 
 from ..externals import _arff
 from ..externals._arff import ArffSparseDataType
-from ..utils import (
-    _chunk_generator,
-    check_pandas_support,
-    get_chunk_n_rows,
-)
+from ..utils import check_pandas_support
+from ..utils._chunking import _chunk_generator, get_chunk_n_rows
 
 
 def _split_sparse_columns(
