+++ sklearn/ensemble/_iforest.py
@@ -12,12 +12,8 @@ from scipy.sparse import issparse
 from ..base import OutlierMixin, _fit_context
 from ..tree import ExtraTreeRegressor
 from ..tree._tree import DTYPE as tree_dtype
-from ..utils import (
-    check_array,
-    check_random_state,
-    gen_batches,
-    get_chunk_n_rows,
-)
+from ..utils import check_array, check_random_state
+from ..utils._chunking import gen_batches, get_chunk_n_rows
 from ..utils._param_validation import Interval, RealNotInt, StrOptions
 from ..utils.validation import _num_samples, check_is_fitted
 from ._bagging import BaseBagging
