+++ sklearn/utils/_encode.py
@@ -4,7 +4,7 @@ from typing import NamedTuple
 
 import numpy as np
 
-from . import is_scalar_nan
+from ._missing import is_scalar_nan
 
 
 def _unique(values, *, return_inverse=False, return_counts=False):
